requirements.txt
pyfiglet
termcolor
colorama
ngram
numpy
flashtext
scikit-learn
nameparser
google-cloud-language
six
scipy
pdfminer.six
PyPDF2
requests
flashtext
nltk
certifi
requests
urllib3
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
